# api/openapi.yaml
openapi: 3.0.3
info:
  title: Data Lake Provisioner API
  version: 0.1.0
  description: |
    MVP Provisioner API for Movincloud Data Lake service.
    Built-in authn/authz is deferred; tenant context is currently supplied via X-Tenant.
servers:
  - url: http://localhost:8081

paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/lakes:
    get:
      summary: List data lakes for the current tenant
      operationId: listLakes
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Tenant lake list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lake'
    post:
      summary: Provision a new data lake
      operationId: provisionLake
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisionLakeRequest'
      responses:
        '202':
          description: Provision operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/stats/summary:
    get:
      summary: Get fleet-wide usage summary across all lakes
      operationId: getFleetUsageSummary
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Fleet usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetUsageSummary'

  /v1/tenant/summary:
    get:
      summary: Get tenant usage summary across all lakes owned by the tenant
      operationId: getTenantUsageSummary
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Tenant usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FleetUsageSummary'

  /v1/lakes/{lakeId}:
    get:
      summary: Get data lake details
      operationId: getLake
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
      responses:
        '200':
          description: Data lake details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lake'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Deprovision data lake
      operationId: deprovisionLake
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '202':
          description: Deprovision operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/lakes/{lakeId}/resize:
    post:
      summary: Resize lake quota
      operationId: resizeLake
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResizeLakeRequest'
      responses:
        '202':
          description: Resize operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/lakes/{lakeId}/operations:
    get:
      summary: List operations for a data lake
      operationId: listLakeOperations
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
      responses:
        '200':
          description: Lake operation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/lakes/{lakeId}/buckets:
    post:
      summary: Create bucket in a lake
      operationId: createBucket
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBucketRequest'
      responses:
        '202':
          description: Bucket create operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      summary: List buckets in a lake
      operationId: listBuckets
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
      responses:
        '200':
          description: Bucket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bucket'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/lakes/{lakeId}/buckets/{bucketId}:
    get:
      summary: Get bucket details
      operationId: getBucket
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
        - $ref: '#/components/parameters/BucketIdPath'
      responses:
        '200':
          description: Bucket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bucket'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete bucket from a lake
      operationId: deleteBucket
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
        - $ref: '#/components/parameters/BucketIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '202':
          description: Bucket delete operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/internal/lakes/{lakeId}/access:
    post:
      summary: Get backend-only temporary internal S3 access for a lake
      description: Backend-only endpoint intended for trusted internal app/server usage, not direct browser or end-user access.
      operationId: getLakeInternalAccess
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
      responses:
        '200':
          description: Backend-only internal lake S3 access lease
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LakeInternalAccess'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/internal/lakes/{lakeId}/customer-s3-access:
    post:
      summary: Ensure and return customer-facing S3 access for a lake
      description: Internal platform endpoint that ensures an active customer S3 key exists for the lake and returns it.
      operationId: getLakeCustomerS3Access
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/LakeIdPath'
      responses:
        '200':
          description: Active customer S3 credential for the lake
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LakeCustomerS3Access'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /v1/operations/{operationId}:
    get:
      summary: Get operation status
      operationId: getOperation
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/OperationIdPath'
      responses:
        '200':
          description: Operation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    TenantHeader:
      in: header
      name: X-Tenant
      required: true
      schema:
        type: string
      description: Tenant context propagated by Movincloud.

    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
      description: Unique key to safely retry create/resize/delete operations.

    LakeIdPath:
      in: path
      name: lakeId
      required: true
      schema:
        type: string

    BucketIdPath:
      in: path
      name: bucketId
      required: true
      schema:
        type: string

    OperationIdPath:
      in: path
      name: operationId
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ProvisionLakeRequest:
      type: object
      required: [sizeGiB]
      properties:
        sizeGiB:
          type: integer
          format: int64
          minimum: 1
          example: 100

    ResizeLakeRequest:
      type: object
      required: [sizeGiB]
      properties:
        sizeGiB:
          type: integer
          format: int64
          minimum: 1
          example: 250

    CreateBucketRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: raw

    Lake:
      type: object
      properties:
        lakeId:
          type: string
        tenantId:
          type: string
        requestedSizeGiB:
          type: integer
          format: int64
        status:
          type: string
          enum: [provisioning, ready, resizing, deleting, failed, deleted]
        rgwUser:
          type: string
        bucketCount:
          type: integer
          format: int32
        usedBytes:
          type: integer
          format: int64
        objectCount:
          type: integer
          format: int64
        lastError:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Bucket:
      type: object
      properties:
        bucketId:
          type: string
        lakeId:
          type: string
        tenantId:
          type: string
        name:
          type: string
        bucketName:
          type: string
        status:
          type: string
          enum: [creating, ready, deleting, failed, deleted]
        usedBytes:
          type: integer
          format: int64
        objectCount:
          type: integer
          format: int64
        lastError:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    FleetUsageSummary:
      type: object
      properties:
        lakeCount:
          type: integer
          format: int32
        bucketCount:
          type: integer
          format: int32
        totalUsedBytes:
          type: integer
          format: int64
        totalCommittedBytes:
          type: integer
          format: int64
        totalObjectCount:
          type: integer
          format: int64

    LakeInternalAccess:
      type: object
      properties:
        rgwUser:
          type: string
        s3Endpoint:
          type: string
        s3Region:
          type: string
        accessKeyId:
          type: string
        secretAccessKey:
          type: string
        leaseExpiresAt:
          type: string
          format: date-time

    LakeCustomerS3Access:
      type: object
      properties:
        lakeId:
          type: string
        rgwUser:
          type: string
        s3Endpoint:
          type: string
        s3Region:
          type: string
        accessKeyId:
          type: string
        secretAccessKey:
          type: string
        issuedAt:
          type: string
          format: date-time
        credentialStatus:
          type: string
          enum: [active]

    Operation:
      type: object
      properties:
        operationId:
          type: string
        operationType:
          type: string
          enum: [provision, resize, deprovision, bucket_create, bucket_delete]
        lakeId:
          type: string
        bucketId:
          type: string
        tenantId:
          type: string
        status:
          type: string
          enum: [pending, running, success, failed]
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string